requirements.txt
streamlit
requests
pandas
pydantic
google-genai
google-cloud-texttospeech
google-auth
google-api-core
mutagen
